[refactor] #168 - facade에서 한 번에 repo를 보지 않도록 service에서 db를 접근하도록 변경
