+++ src/main/java/gitbal/backend/domain/guestbook/application/GuestBookService.java
@@ -1,7 +1,7 @@
 package gitbal.backend.domain.guestbook.application;
 
-import gitbal.backend.api.dashboardPage.dto.GuestBookResponseDto;
-import gitbal.backend.api.dashboardPage.dto.GuestBookWriteRequestDto;
+import gitbal.backend.api.guestBookPage.dto.GuestBookResponseDto;
+import gitbal.backend.api.guestBookPage.dto.GuestBookWriteRequestDto;
 import gitbal.backend.domain.user.User;
 import java.util.List;
