+++ src/main/java/gitbal/backend/domain/guestbook/application/GuestBookServiceImpl.java
@@ -1,8 +1,8 @@
 package gitbal.backend.domain.guestbook.application;
 
 
-import gitbal.backend.api.dashboardPage.dto.GuestBookResponseDto;
-import gitbal.backend.api.dashboardPage.dto.GuestBookWriteRequestDto;
+import gitbal.backend.api.guestBookPage.dto.GuestBookResponseDto;
+import gitbal.backend.api.guestBookPage.dto.GuestBookWriteRequestDto;
 import gitbal.backend.domain.guestbook.GuestBook;
 import gitbal.backend.domain.guestbook.infra.repository.GuestBookRepository;
 import gitbal.backend.domain.user.User;
